add some functions to work with db. delete expired rows + isNullOrEmpty table + isActive

+++ sdk/src/main/java/org/tapsell/sdk/data/datasource/WaterfallDataSource.kt
@@ -1,11 +1,10 @@
 package org.tapsell.sdk.data.datasource
 
-import android.util.Log
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.onEach
-import org.tapsell.sdk.data.local.db.user.WaterfallEntity
+import org.tapsell.sdk.addmediator.extensions.isNullOrEmpty
+import org.tapsell.sdk.data.local.db.waterfall.WaterfallEntity
 import org.tapsell.sdk.data.model.response.WaterfallResponse
 import org.tapsell.sdk.data.model.response.toLocalEntity
 import org.tapsell.sdk.data.repository.LocalWaterfallRepositoryImpl
@@ -24,17 +23,17 @@ internal object WaterfallDataSource {
 
     suspend fun getActiveWaterfallOrNull(): WaterfallEntity? {
         var activeWaterfall: WaterfallEntity? = null
-        readAllRecords().onEach { waterfall: WaterfallEntity ->
-            if (isWaterfallExpired(waterfall.timestamp)) {
-                // delete from db if expired
-                deleteRecord(waterfall)
-            } else {
-                activeWaterfall = waterfall
-                return@onEach
+        readAllRecords().filterNotNull().onEach { waterfalls: List<WaterfallEntity> ->
+            waterfalls.forEach { waterfall: WaterfallEntity ->
+                if (isWaterfallExpired(waterfall.timestamp)) {
+                    // delete from db if expired
+                    deleteRecord(waterfall)
+                } else {
+                    activeWaterfall = waterfall
+                    return@onEach
+                }
             }
-        }.catch { e: Throwable ->
-            Log.e("Error: $TAG :", e.message?.toString() ?: "")
-        }.collect()
+        }
         return activeWaterfall
     }
 
@@ -46,27 +45,29 @@ internal object WaterfallDataSource {
         LocalWaterfallRepositoryImpl.insertAll(flowList.toLocalEntity())
     }
 
-    private suspend fun readAllRecords(): Flow<WaterfallEntity> {
+    private suspend fun readAllRecords(): Flow<List<WaterfallEntity>> {
         return LocalWaterfallRepositoryImpl.getAll()
     }
 
-    suspend fun getRecord(id: String): WaterfallEntity? {
-        return LocalWaterfallRepositoryImpl.getItem(id)
-    }
-
-    suspend fun store(waterfall: WaterfallResponse) {
-        LocalWaterfallRepositoryImpl.insert(waterfall.toLocalEntity())
-    }
-
-    suspend fun update(waterfall: WaterfallEntity) {
-        LocalWaterfallRepositoryImpl.update(waterfall)
-    }
-
-    suspend fun deleteRecord(waterfall: WaterfallEntity) {
+    private suspend fun deleteRecord(waterfall: WaterfallEntity) {
         return LocalWaterfallRepositoryImpl.delete(waterfall)
     }
 
-    suspend fun deleteAllRecords() {
-        return LocalWaterfallRepositoryImpl.deleteAll()
+    suspend fun deleteAllExpiredRecords() {
+        if (isWaterfallTableNullOrEmpty()) {
+            return
+        }
+        LocalWaterfallRepositoryImpl.deleteAll()
+    }
+
+    private suspend fun isWaterfallTableNullOrEmpty(): Boolean {
+        return LocalWaterfallRepositoryImpl
+            .getAll()
+            .isNullOrEmpty()
+    }
+
+
+    suspend fun isActiveWaterfall(): Boolean {
+        return getActiveWaterfallOrNull() !== null
     }
 }
